hoist Future over typeinf_local
Uses a slightly different approach for the Future, since we just have
one of these cases, it doesn't need to go into a generalized queue. So,
instead of a general closure, return a specific struct for the resume
state which can be held in a queue in the caller.

It is slightly awkward that we have both CurrentState and InferenceState
which contain a random split of variables between the two. But that is
effectively what we had before, just on the stack.